requirements.txt
astroid==2.2.5
backcall==0.1.0
Click==7.0
decorator==4.4.0
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.1
PyJWT==1.7.1
pylint==2.3.1
pylint-flask==0.6
pylint-plugin-utils==0.5
six==1.12.0
SQLAlchemy==1.3.3
traitlets==4.3.2
typed-ast==1.3.5
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.1
